Edited RemoveUser to work with userSchema, and prevent delete of admin

+++ backend/src/app/services/removeUser.service.ts
@@ -8,20 +8,18 @@ import Pharmacist from '../schemas/pharmacist';
 export async function removeUser(username: string): Promise<void> {
 
     const user = await User.findOne({ username });
-    const pharm = await Pharmacist.findOne({ username });
   
       console.log(user);
-      console.log(pharm);
       if(user){
+        if (user.role == "ADMINISTRATOR"){
+          throw new AppError("cannot delete an admin", 406, ERROR);
+        }
         await Patient.findOneAndDelete({ user: user._id });
+        await Pharmacist.findOneAndDelete({ user: user._id });
         await User.findOneAndDelete({ username });
-      }else  if (pharm) {
-      await Pharmacist.findOneAndDelete({ username });
-     
+        return
     }
-    else{
+   
       throw new AppError("This user doesn't exist", 404, ERROR);
       
     }
-
-}
